Finalização da pesquisa e outras correções finais

+++ interface/apps/main/models.py
@@ -79,6 +79,6 @@ class Solicitacao(models.Model):
 class Recursos(models.Model):
     semestre = models.CharField(max_length=6, primary_key=True)
     valor_disponivel = models.FloatField()
-    porcentagem_gp1 = models.IntegerField()
-    porcentagem_gp2 = models.IntegerField()
-    porcentagem_gp3 = models.IntegerField()
+    porcentagem_gp1 = models.FloatField()
+    porcentagem_gp2 = models.FloatField()
+    porcentagem_gp3 = models.FloatField()
